requirements.txt
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.7
langchain-community>=0.2.0
chromadb>=0.5.0
pypdf>=4.2.0
requests>=2.31.0
python-dotenv>=1.0.1
tiktoken>=0.7.0